Random Picker: fold the sound feature back into the tool file
Keep the diff against main minimal: the wheel/UI code returns to its
original place untouched, and the branch now reads as one small
addition — the WebAudio tick/ding, the rAF slice-crossing loop, and
the mute toggle. audio.ts / useSpinWheel.ts / Wheel.tsx are gone.
